Add build and test on pull request workflow
This commit adds a new workflow file, ".github/workflows/build-test-pull-request.yml", which sets up a job to build and test the code on pull requests. This workflow runs on the "ubuntu-latest" runner and includes steps to checkout the code, install dependencies, build all services, and run unit tests. The commit also includes updates to the "docker-compose.prod.yml" file, where the image tags for the services are now set to "${COMPOSE_TAG:-latest}".